Fix compiler warnings

+++ task2/src/cuckoolock.hpp
@@ -37,7 +37,7 @@ public:
 
     void quiesce()
     {
-        for(int i = 0; i < the_capacity; i++) {
+        for(size_t i = 0; i < the_capacity; i++) {
             the_lock[0][i].lock();
             the_lock[0][i].unlock();
         }
